we dont use aac layers now, so refactor code and comment out
