+++ components/map/layer-manager/component.js
@@ -4,9 +4,8 @@ import PropTypes from 'prop-types';
 import { LayerManager, Layer } from 'layer-manager/dist/components';
 import { PluginMapboxGl, fetch } from 'layer-manager';
 
-import { omit } from 'utils/general';
-
-import slugify from 'slugify';
+// import fmuAccProvider from './providers/fmu-aac-provider';
+import countriesProvider from './providers/countries-provider';
 
 class LayerManagerComponent extends PureComponent {
   static propTypes = {
@@ -22,117 +21,10 @@ class LayerManagerComponent extends PureComponent {
         map={map}
         plugin={PluginMapboxGl}
         providers={{
-          countries: (layerModel, layer, resolve, reject) => {
-            const { source } = layerModel;
-            const { provider } = source;
-
-            fetch('get', provider.url, provider.options, layerModel)
-              .then(({ data }) => {
-                return resolve({
-                  ...layer,
-                  source: {
-                    ...omit(layer.source, 'provider'),
-                    data: data.attributes.geojson
-                  }
-                });
-              })
-              .catch(e => {
-                reject(e);
-              });
-          },
-          'aac-cog': (layerModel, layer, resolve, reject) => {
-            const { source } = layerModel;
-            const { provider } = source;
-
-            fetch('get', provider.url, provider.options, layerModel)
-              .then((data) => {
-                const parsedData = {
-                  ...data,
-                  features: data.features.map(f => ({
-                    ...f,
-                    properties: {
-                      ...f.properties,
-                      num_con: slugify(f.properties.num_con, {
-                        lower: true
-                      })
-                    }
-                  }))
-                };
-
-                return resolve({
-                  ...layer,
-                  source: {
-                    ...omit(layer.source, 'provider'),
-                    data: parsedData
-                  }
-                });
-              })
-              .catch(e => {
-                reject(e);
-              });
-          },
-          'aac-cod': (layerModel, layer, resolve, reject) => {
-            const { source } = layerModel;
-            const { provider } = source;
-
-            fetch('get', provider.url, provider.options, layerModel)
-              .then((data) => {
-                const parsedData = {
-                  ...data,
-                  features: data.features.map(f => ({
-                    ...f,
-                    properties: {
-                      ...f.properties,
-                      num_ccf: slugify(f.properties.num_ccf || '', {
-                        lower: true
-                      })
-                    }
-                  }))
-                };
-
-                return resolve({
-                  ...layer,
-                  source: {
-                    ...omit(layer.source, 'provider'),
-                    data: parsedData
-                  }
-                });
-              })
-              .catch(e => {
-                reject(e);
-              });
-          },
-          'aac-cmr': (layerModel, layer, resolve, reject) => {
-            const { source } = layerModel;
-            const { provider } = source;
-
-            fetch('get', provider.url, provider.options, layerModel)
-              .then((data) => {
-                const parsedData = {
-                  ...data,
-                  features: data.features.map(f => ({
-                    ...f,
-                    properties: {
-                      ...f.properties,
-                      nom_ufe: slugify(f.properties.nom_ufe || '', {
-                        lower: true
-                      })
-                    }
-                  }))
-                };
-
-                return resolve({
-                  ...layer,
-                  source: {
-                    ...omit(layer.source, 'provider'),
-                    data: parsedData
-                  }
-                });
-              })
-              .catch(e => {
-                reject(e);
-              });
-          }
+          countries: countriesProvider,
+          // 'aac-cog': fmuAccProvider('num_con'),
+          // 'aac-cod': fmuAccProvider('num_ccf'),
+          // 'aac-cmr': fmuAccProvider('nom_ufe')
         }}
       >
         {!!layers && layers.map((l) => {
