+++ constants/layers.js
@@ -320,112 +320,112 @@ export const LAYERS = [
       source: 'layer.tree-cover-loss.metadata.source'
     }
   },
-  {
-    id: 'aac-cog',
-    name: 'aac',
-    iso: 'COG',
-    config: {
-      type: 'geojson',
-      source: {
-        type: 'geojson',
-        provider: {
-          type: 'aac-cog',
-          url: 'https://opendata.arcgis.com/datasets/0c31460808d84dfe806127a25fd0de62_29.geojson'
-        }
-      },
-      render: {
-        layers: [
-          {
-            type: 'fill',
-            paint: {
-              'fill-color': '#CCCCCC',
-              'fill-opacity': 0.5
-            },
-            filter: [
-              'all',
-              ['in', ['get', 'num_con'], ['literal', '{fmuNames}']]
-            ]
-          },
-          {
-            type: 'line',
-            paint: {
-              'line-color': '#CCCCCC',
-              'line-opacity': 0.5
-            },
-            filter: [
-              'all',
-              ['in', ['get', 'num_con'], ['literal', '{fmuNames}']]
-            ]
-          }
-        ]
-      }
-    },
-    paramsConfig: [
-      { key: 'fmuNames', default: [], required: true }
-    ],
-    legendConfig: {
-      type: 'basic',
-      items: [
-        { name: 'aac', color: '#CCCCCC' }
-      ]
-    },
-    interactionConfig: {
-      enabled: true
-    }
-  },
-  {
-    id: 'aac-cod',
-    name: 'aac',
-    iso: 'COD',
-    config: {
-      type: 'geojson',
-      source: {
-        type: 'geojson',
-        provider: {
-          type: 'aac-cod',
-          url: 'https://opendata.arcgis.com/datasets/c60d5bf9e01c45c5ad79208803819db1_30.geojson'
-        }
-      },
-      render: {
-        layers: [
-          {
-            type: 'fill',
-            paint: {
-              'fill-color': '#CCCCCC',
-              'fill-opacity': 0.5
-            },
-            filter: [
-              'all',
-              ['in', ['get', 'num_ccf'], ['literal', '{fmuNames}']]
-            ]
-          },
-          {
-            type: 'line',
-            paint: {
-              'line-color': '#CCCCCC',
-              'line-opacity': 0.5
-            },
-            filter: [
-              'all',
-              ['in', ['get', 'num_ccf'], ['literal', '{fmuNames}']]
-            ]
-          }
-        ]
-      }
-    },
-    paramsConfig: [
-      { key: 'fmuNames', default: [], required: true }
-    ],
-    legendConfig: {
-      type: 'basic',
-      items: [
-        { name: 'aac', color: '#CCCCCC' }
-      ]
-    },
-    interactionConfig: {
-      enabled: true
-    }
-  },
+  // {
+  //   id: 'aac-cog',
+  //   name: 'aac',
+  //   iso: 'COG',
+  //   config: {
+  //     type: 'geojson',
+  //     source: {
+  //       type: 'geojson',
+  //       provider: {
+  //         type: 'aac-cog',
+  //         url: 'https://opendata.arcgis.com/datasets/0c31460808d84dfe806127a25fd0de62_29.geojson'
+  //       }
+  //     },
+  //     render: {
+  //       layers: [
+  //         {
+  //           type: 'fill',
+  //           paint: {
+  //             'fill-color': '#CCCCCC',
+  //             'fill-opacity': 0.5
+  //           },
+  //           filter: [
+  //             'all',
+  //             ['in', ['get', 'num_con'], ['literal', '{fmuNames}']]
+  //           ]
+  //         },
+  //         {
+  //           type: 'line',
+  //           paint: {
+  //             'line-color': '#CCCCCC',
+  //             'line-opacity': 0.5
+  //           },
+  //           filter: [
+  //             'all',
+  //             ['in', ['get', 'num_con'], ['literal', '{fmuNames}']]
+  //           ]
+  //         }
+  //       ]
+  //     }
+  //   },
+  //   paramsConfig: [
+  //     { key: 'fmuNames', default: [], required: true }
+  //   ],
+  //   legendConfig: {
+  //     type: 'basic',
+  //     items: [
+  //       { name: 'aac', color: '#CCCCCC' }
+  //     ]
+  //   },
+  //   interactionConfig: {
+  //     enabled: true
+  //   }
+  // },
+  // {
+  //   id: 'aac-cod',
+  //   name: 'aac',
+  //   iso: 'COD',
+  //   config: {
+  //     type: 'geojson',
+  //     source: {
+  //       type: 'geojson',
+  //       provider: {
+  //         type: 'aac-cod',
+  //         url: 'https://opendata.arcgis.com/datasets/c60d5bf9e01c45c5ad79208803819db1_30.geojson'
+  //       }
+  //     },
+  //     render: {
+  //       layers: [
+  //         {
+  //           type: 'fill',
+  //           paint: {
+  //             'fill-color': '#CCCCCC',
+  //             'fill-opacity': 0.5
+  //           },
+  //           filter: [
+  //             'all',
+  //             ['in', ['get', 'num_ccf'], ['literal', '{fmuNames}']]
+  //           ]
+  //         },
+  //         {
+  //           type: 'line',
+  //           paint: {
+  //             'line-color': '#CCCCCC',
+  //             'line-opacity': 0.5
+  //           },
+  //           filter: [
+  //             'all',
+  //             ['in', ['get', 'num_ccf'], ['literal', '{fmuNames}']]
+  //           ]
+  //         }
+  //       ]
+  //     }
+  //   },
+  //   paramsConfig: [
+  //     { key: 'fmuNames', default: [], required: true }
+  //   ],
+  //   legendConfig: {
+  //     type: 'basic',
+  //     items: [
+  //       { name: 'aac', color: '#CCCCCC' }
+  //     ]
+  //   },
+  //   interactionConfig: {
+  //     enabled: true
+  //   }
+  // },
   // {
   //   id: 'aac-cmr',
   //   name: 'aac',
